+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 0.0.1
 
 - Initial version.
+
+## 0.0.2
+
+- Added compatibility for Dataclass-Engine 0.0.10 (the toApi release)
+- Added removeTypeKey function